docs: ADHD - How to ADHD vid notes
I want to not stress so much about projects and was trying to think of
the why behind it.
Thought it had to do with how I kind-of expect rejection and stress out
about it a lot.
I started off trying to look into Rejection Sensitivity Dysmorphia
Afterwards into project planning stuff